requirements.txt
google-genai
gspread
streamlit
gspread_formatting
dotenv